tidy up types in separate file
